creating controls to show on card hover
