+++ README.md
@@ -1,1 +1,2 @@
 "If you can't do what you need to, do wutchyacan."
+current version: v0.1.0
