+++ src/components/Card/InlineCard/inlineCard.module.scss
