+++ src/components/Card/InlineControlsOverlay/InlineControlsOverlay.tsx
@@ -1,0 +1,11 @@
+import MoreHorizIcon from '@mui/icons-material/MoreHoriz';
+
+const InlineControlsOverlay = () => {
+    return (
+        <div>
+            <MoreHorizIcon />
+        </div>
+    )
+}
+
+export default InlineControlsOverlay;
